Code improvements

+++ app.py
@@ -1,7 +1,7 @@
 import os
 from flask import (
-    Flask, render_template, redirect, 
-    request, session, url_for, flash
+    Flask, render_template, redirect,
+    request, session, url_for, flash, abort
 )
 from flask_pymongo import PyMongo
 from bson.objectid import ObjectId
@@ -34,38 +34,42 @@ def home():
 @app.route("/browse-recipes")
 def browse_recipes():
     recipes = mongo.db.recipes.find()
-    return render_template("browse-recipes.html", recipes=recipes)     
+    return render_template("browse-recipes.html", recipes=recipes)
 
 
 # search recipes
 @app.route("/search_recipe", methods=["GET", "POST"])
 def search_recipe():
     category_select = request.form.get("category_select")
-    ingredient_search = request.form.get("ingredient_search") 
+    ingredient_search = request.form.get("ingredient_search")
 
-    # loop creating different search options depending on what has been populated
+    # loop creating different search options
     if category_select == "all-types":
         if ingredient_search:
-        # if all food types are selected and the text input is populated, search for the specific word
-            recipe_search = list(mongo.db.recipes.find({"$text": {"$search": ingredient_search}}))
+            # if all food types are selected and the text input is populated
+            # search for the specific word
+            recipe_search = list(mongo.db.recipes.find(
+                {"$text": {"$search": ingredient_search}}))
         else:
-        #search for everything if nothing is specified in the search
-            recipe_search = mongo.db.recipes.find()   
+            # search for everything if nothing is specified in the search
+            recipe_search = mongo.db.recipes.find()
     elif category_select == "my_recipes":
-        #search for user specific recipes and text search
         if ingredient_search:
+            # search for user specific recipes and text search
             recipe_search = list(mongo.db.recipes.find(
                 {"$and": [
                     {"created_by": session["user"]},
                     {"$text": {"$search": ingredient_search}}
                 ]}
             ))
-        # search only for user-specific recipes
         else:
-            recipe_search = list(mongo.db.recipes.find({"created_by": session["user"]}))
+            # search only for user-specific recipes
+            recipe_search = list(mongo.db.recipes.find(
+                {"created_by": session["user"]}))
     else:
         if ingredient_search:
-        # if both dropdown and input area are populated, search for both 
+            # if both dropdown and input area are populated
+            # search for both
             recipe_search = list(mongo.db.recipes.find(
                 {"$and": [
                     {"food_category": category_select},
@@ -73,32 +77,38 @@ def search_recipe():
                 ]}
             ))
         else:
-        # if only a food category is selected, search by that
+            # if only a food category is selected, search by that
             recipe_search = list(mongo.db.recipes.find(
-                {"food_category": category_select})) 
+                {"food_category": category_select}))
 
-    return render_template("browse-recipes.html", recipes=recipe_search) 
+    return render_template("browse-recipes.html", recipes=recipe_search)
 
 
 # my recipes menu option, searches for "my recipes" in browse-recipes
 @app.route("/my_recipes", methods=["GET", "POST"])
 def my_recipes():
-    recipe_search = list(mongo.db.recipes.find({"created_by": session["user"]}))
-    return render_template("browse-recipes.html", recipes=recipe_search) 
+    if "user" in session:
+        recipe_search = list(mongo.db.recipes.find(
+            {"created_by": session["user"]}))
+        return render_template("browse-recipes.html", recipes=recipe_search)
+    abort(403)
+
 
 # view recipe
 @app.route("/view-recipe/<recipe_id>")
 def view_recipe(recipe_id):
     recipe = mongo.db.recipes.find_one({"_id": ObjectId(recipe_id)})
-    return render_template("view-recipe.html", recipe=recipe)
+    if recipe:
+        return render_template("view-recipe.html", recipe=recipe)
+    abort(404)
 
 
 # register
 @app.route("/register", methods=["GET", "POST"])
 def register():
-    # TODO: change the redirect to an 403 error page 
+    # TODO: change the redirect to an 403 error page
     if "user" in session:
-        return redirect(url_for("home"))
+        abort(403)
 
     if request.method == "POST":
         existing_user = mongo.db.users.find_one(
@@ -108,13 +118,13 @@ def register():
             flash("Name already taken!")
             return redirect(url_for("register"))
 
-        register = {
+        user_data = {
             "username": request.form.get("username").lower(),
             "password": generate_password_hash(request.form.get("password"))
-        }   
-        mongo.db.users.insert_one(register)     
+        }
+        mongo.db.users.insert_one(user_data)
 
-        session["user"] = request.form.get("username").lower() 
+        session["user"] = request.form.get("username").lower()
         flash("Successfully Registered!")
         return redirect(url_for("home"))
 
@@ -124,27 +134,27 @@ def register():
 # log in
 @app.route("/login", methods=["GET", "POST"])
 def login():
-    # TODO: change the redirect to an 403 error page 
+    # TODO: change the redirect to an 403 error page
     if "user" in session:
-        return redirect(url_for("home"))
-        
+        abort(403)
+
     if request.method == "POST":
         # if username exists in the database
         existing_user = mongo.db.users.find_one(
             {"username": request.form.get("username").lower()})
-        
-        if existing_user:  
-            # correct passord
+
+        if existing_user:
             if check_password_hash(
-                existing_user["password"], request.form.get("password")):
+                    existing_user["password"], request.form.get("password")):
+                # correct passord
                 session["user"] = request.form.get("username").lower()
                 return redirect(url_for("home"))
-            # incorrect password
             else:
+                # incorrect password
                 flash("Incorrect password or username, try again!")
                 return redirect(url_for("login"))
-        # if username doesn't exist
         else:
+            # if username doesn't exist
             flash("Incorrect password or username, try again!")
             return redirect(url_for("login"))
 
@@ -154,65 +164,94 @@ def login():
 # log out
 @app.route("/logout")
 def logout():
-    # TODO: change the redirect to an 403 error page 
+    # TODO: change the redirect to an 403 error page
     if "user" in session:
-
         session.pop("user")
         flash("See you later!")
+
         return redirect(url_for("login"))
-        
-    return redirect(url_for("home"))        
+
+    abort(403)
 
 
 # delete account
 @app.route("/delete_account")
 def delete_account():
-    # TODO: change the redirect to an 403 error page 
+    # TODO: change the redirect to an 403 error page
     if "user" in session:
-
         user_id = mongo.db.users.find_one({"username": session["user"]})["_id"]
         mongo.db.users.delete_one({"_id": ObjectId(user_id)})
         session.pop("user")
         flash("Sad to see you go, you can regster with us again at any time!!")
+
         return redirect(url_for("register"))
 
-    return redirect(url_for("home"))    
+    abort(403)
 
 
 # create recipe
 @app.route("/create-recipe", methods=["GET", "POST"])
 def create_recipe():
-    # TODO: change the #redirect to an 403 error page 
+    # TODO: change the #redirect to an 403 error page
     if "user" in session:
         if request.method == "POST":
             recipe = json.loads(request.get_data(as_text=True))
             recipe["created_by"] = session["user"]
             mongo.db.recipes.insert_one(recipe)
-        
+
         return render_template("create-recipe.html", recipe=0)
 
-    return redirect(url_for("home"))
+    abort(403)
 
 
 # edit recipe
 @app.route("/edit_recipe/<recipe_id>", methods=["GET", "POST"])
 def edit_recipe(recipe_id):
-    if request.method == "POST":
-        recipe = json.loads(request.get_data(as_text=True))
-        recipe["created_by"] = session["user"]
-        mongo.db.recipes.update_one({"_id": ObjectId(recipe_id)}, {"$set": recipe})
-        flash("Recipe successfully updated")
-        recipes = mongo.db.recipes.find()
-        return render_template("home.html", recipes=recipes)
-    recipe = mongo.db.recipes.find_one({"_id": ObjectId(recipe_id)})
-    return render_template("create-recipe.html", recipe=recipe)
+    if "user" in session:
+        recipe = mongo.db.recipes.find_one({"_id": ObjectId(recipe_id)})
+        if session["user"] == recipe["created_by"]:
+            if request.method == "POST":
+                recipe = json.loads(request.get_data(as_text=True))
+                recipe["created_by"] = session["user"]
+                mongo.db.recipes.update_one(
+                    {"_id": ObjectId(recipe_id)}, {"$set": recipe})
+                flash("Recipe successfully updated")
+                recipes = mongo.db.recipes.find()
+                return render_template("home.html", recipes=recipes)
+
+            return render_template("create-recipe.html", recipe=recipe)
+
+    abort(403)
 
 
 # delete recipe
 @app.route("/delete_recipe/<recipe_id>", methods=["GET", "POST"])
 def delete_recipe(recipe_id):
-    mongo.db.recipes.delete_one({"_id": ObjectId(recipe_id)})
-    return redirect(url_for("home"))  
+    if "user" in session:
+        recipe = mongo.db.recipes.find_one({"_id": ObjectId(recipe_id)})
+        if session["user"] == recipe.created_by:
+            mongo.db.recipes.delete_one({"_id": ObjectId(recipe_id)})
+
+            return redirect(url_for("home"))
+
+    abort(403)
+
+
+# error functions below
+
+@app.errorhandler(403)
+def page_forbidden(e):
+    return render_template("403.html"), 403
+
+
+@app.errorhandler(404)
+def page_not_found(e):
+    return render_template("404.html"), 404
+
+
+@app.errorhandler(500)
+def internal_server_error(e):
+    return render_template("500.html"), 500
 
 
 if __name__ == "__main__":
